server/requirements.txt
# FastAPI 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
# pgvector==0.2.4  # 已移除，使用ChromaDB替代

# 异步任务队列
celery==5.3.4
redis==5.0.1

# 文件存储
minio==7.2.0
Pillow==10.1.0

# AI/ML 模型
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
sentence-transformers==2.2.2
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.2

# 图像处理
Pillow==10.1.0
opencv-python==4.8.1.78
exifread==3.0.0

# 向量数据库
chromadb==0.4.18

# 工具库
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
